highScore issue
