+++ src/components/App.js
@@ -1,90 +1,21 @@
-import React, { Component } from 'react';
-import {Provider} from './Context';
+import React from 'react';
 import Header from './Header';
 import PlayerList from './PlayerList';
 import AddPlayerForm from './AddPlayerForm';
 
-class App extends Component {
-  state = {
-    players: [
-      {
-        name: "Player 1",
-        score: 0,
-        id: 1
-      },
-      {
-        name: "Player 2",
-        score: 0,
-        id: 2
-      },
-    ]
-  };
+const App = () => {
+  const highScore = this.getHighScore();
 
-  //player id counter
-  prevPlayerId = 4;
+  return (
+    <div className="scoreboard">
+      <Header />
+        <PlayerList
+          highScore={highScore}         
+        />
 
-  getHighScore = () => {
-    const scores = this.state.players.map( p => p.score );
-    const highScore = Math.max(...scores);
-    if (highScore) {
-      return highScore;
-    } 
-    return null;
-  }
-
-  handleScoreChange = (index, delta) => {
-    this.setState( prevState => ({
-      score: prevState.players[index].score += delta
-    }));
-  }
-
-  handleAddPlayer = (name) => {
-    this.setState( prevState => {
-      return {
-        players: [
-          ...prevState.players,
-          {
-            name,
-            score: 0,
-            id: this.prevPlayerId += 1
-          }
-        ]
-      };
-    });
-  }
-
-  handleRemovePlayer = (id) => {
-    this.setState( prevState => {
-      return {
-        players: prevState.players.filter(p => p.id !== id)
-      };
-    });
-  }
-
-  render() {
-    const highScore = this.getHighScore();
-
-    return (
-      <Provider value={{
-        players: this.state.players,
-        actions: {
-          changeScore: this.handleScoreChange,
-          removePlayer: this.handleRemovePlayer,
-          addPlayer: this.handleAddPlayer
-        },
-        highScore: highScore
-      }}>
-        <div className="scoreboard">
-          <Header />
-            <PlayerList
-              highScore={highScore}         
-            />
-
-          <AddPlayerForm />
-        </div>
-      </Provider>
-    );
-  }
+      <AddPlayerForm />
+    </div>
+  );
 }
 
 export default App;
