+++ src/index.js
@@ -1,10 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
+import {Provider} from './components/Context';
 import App from './components/App';
 import './index.css';
 
 ReactDOM.render(
-  <App />, 
+  <Provider>
+    <App />
+  </Provider>, 
   document.getElementById('root')
 );
